emoji layer: fix uint8_t overflow in category count (cat 11 truncated)
emj_category_t.count and the EMJ_CAT_ENTRY cast were uint8_t. Category
11 (Latin Extended-A & B) has 335 codepoints; 335 & 0xFF = 79, so the
firmware saw only 79 entries and rendered just 0x0100-0x014E — the rest
of Latin Extended-A (0x014F-0x017E) and all of Latin Extended-B
(0x0180-0x024F) were treated as empty slots.

Widen count to uint16_t (struct field, EMJ_CAT_ENTRY cast, and the
local in page_count). Now cat 11 reports 335 entries → 7 pages, all
glyphs reachable.

Verified: split72:default compiles and links cleanly (.uf2 produced).

+++ keyboards/handwired/polykybd/emoji/emoji_data.h
@@ -9,7 +9,7 @@
 
 typedef struct {
     const uint32_t *codepoints;
-    uint8_t         count;
+    uint16_t        count;
 } emj_category_t;
 
 // ── Category 0: Smileys & Faces ─────────────────────────────────────────────
@@ -410,7 +410,7 @@ static const uint32_t emj_cat11[] = {
 };
 
 // ── Category table ──────────────────────────────────────────────────────────
-#define EMJ_CAT_ENTRY(arr) { arr, (uint8_t)(sizeof(arr) / sizeof(arr[0])) }
+#define EMJ_CAT_ENTRY(arr) { arr, (uint16_t)(sizeof(arr) / sizeof(arr[0])) }
 
 static const emj_category_t EMJ_CATEGORIES[] = {
     EMJ_CAT_ENTRY(emj_cat0),   // 0: Smileys    😀
